requirements.txt
Django==1.10.3
djangorestframework==3.6.4
mysqlclient==1.3.12
pytz==2017.3
